DTNN-1258 Implement Duct Flow in PI

+++ bpa/stor/inc/bplib_stor.h
@@ -144,7 +144,7 @@ BPLib_Status_t BPLib_STOR_ScanCache(BPLib_Instance_t* Inst);
 
 BPLib_Status_t BPLib_STOR_StoreBundle(BPLib_Instance_t* Inst, BPLib_Bundle_t* Bundle);
 
-BPLib_Status_t BPLib_STOR_EgressForID(BPLib_Instance_t* Inst, uint16_t EgressID, bool LocalDelivery, 
+BPLib_Status_t BPLib_STOR_EgressForID(BPLib_Instance_t* Inst, uint32_t EgressID, bool LocalDelivery, 
     size_t* NumEgressed);
 
 BPLib_Status_t BPLib_STOR_GarbageCollect(BPLib_Instance_t* Inst, size_t* NumDiscarded);
